requirements.txt
aerich==0.6.3
aiofiles==22.1.0
aiogram @ https://github.com/aiogram/aiogram/archive/refs/heads/dev-3.x.zip
aiogram3-form==0.0.9
aiohttp==3.8.3
aioschedule==0.5.2
aiosignal==1.2.0
aiosqlite==0.17.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.9.24
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
dictdiffer==0.9.0
frozenlist==1.3.0
idna==3.3
iso8601==1.0.2
magic-filter==1.0.9
multidict==6.0.2
pydantic==1.10.2
Pygments==2.12.0
pypika-tortoise==0.1.5
python-dotenv==0.21.0
pytz==2022.1
tomlkit==0.11.4
tortoise-orm==0.19.1
typer==0.4.1
typing_extensions==4.2.0
yarl==1.7.2
